Remove superstore
