+++ settings.gradle
@@ -1,4 +1,3 @@
 include ':store'
-include ':superstore'
 include ':cache'
 include ':multicast'
