--- superstore/build.gradle.kts
@@ -1,130 +1,0 @@
-@file:Suppress("UnstableApiUsage")
-
-import com.vanniktech.maven.publish.SonatypeHost.S01
-import org.jetbrains.dokka.gradle.DokkaTask
-
-plugins {
-    kotlin("multiplatform")
-    kotlin("plugin.serialization")
-    id("com.android.library")
-    id("com.vanniktech.maven.publish")
-    id("org.jetbrains.dokka")
-    id("org.jetbrains.kotlinx.kover")
-    `maven-publish`
-    kotlin("native.cocoapods")
-    id("kotlinx-atomicfu")
-}
-
-kotlin {
-    android()
-    jvm()
-    iosArm64()
-    iosX64()
-    linuxX64()
-    iosSimulatorArm64()
-    js {
-        browser()
-        nodejs()
-    }
-    cocoapods {
-        summary = "Superstore"
-        homepage = "https://github.com/MobileNativeFoundation/Store"
-        ios.deploymentTarget = "13"
-        version = Version.store
-    }
-
-    sourceSets {
-        all {
-            languageSettings.apply {
-                optIn("kotlinx.coroutines.ExperimentalCoroutinesApi")
-                optIn("kotlin.RequiresOptIn")
-            }
-        }
-
-        val commonMain by getting {
-            dependencies {
-                with(Deps.Kotlinx) {
-                    api(atomicFu)
-                    implementation(stdLib)
-                    implementation(serializationCore)
-                    implementation(serializationJson)
-                    implementation(coroutinesCore)
-                }
-                api(project(":store"))
-                implementation(project(":multicast"))
-                implementation(project(":cache"))
-            }
-        }
-
-        val commonTest by getting {
-            dependsOn(commonMain)
-            dependencies {
-                implementation(kotlin("test"))
-                with(Deps.Test) {
-                    implementation(junit)
-                    implementation(core)
-                    implementation(coroutinesTest)
-                }
-            }
-        }
-
-        val jvmMain by getting
-        val androidMain by getting
-        val nativeMain by creating {
-            dependsOn(commonMain)
-        }
-    }
-}
-
-android {
-    sourceSets["main"].manifest.srcFile("src/androidMain/AndroidManifest.xml")
-    compileSdk = 33
-
-    defaultConfig {
-        minSdk = 24
-        targetSdk = 33
-    }
-
-    lint {
-        disable += "ComposableModifierFactory"
-        disable += "ModifierFactoryExtensionFunction"
-        disable += "ModifierFactoryReturnType"
-        disable += "ModifierFactoryUnreferencedReceiver"
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-}
-
-tasks.withType<DokkaTask>().configureEach {
-    dokkaSourceSets.configureEach {
-        reportUndocumented.set(false)
-        skipDeprecated.set(true)
-        jdkVersion.set(8)
-    }
-}
-
-mavenPublishing {
-    publishToMavenCentral(S01)
-    signAllPublications()
-}
-
-koverMerged {
-    enable()
-
-    xmlReport {
-        onCheck.set(true)
-        reportFile.set(layout.projectDirectory.file("kover/coverage.xml"))
-    }
-
-    htmlReport {
-        onCheck.set(true)
-        reportDir.set(layout.projectDirectory.dir("kover/html"))
-    }
-
-    verify {
-        onCheck.set(true)
-    }
-}
