--- superstore/gradle.properties
@@ -1,3 +1,0 @@
-POM_NAME=org.mobilenativefoundation.store
-POM_ARTIFACT_ID=superstore
-POM_PACKAGING=jar
