--- superstore/src/commonMain/kotlin/org/mobilenativefoundation/store/superstore5/SuperstoreResponseOrigin.kt
@@ -1,8 +1,0 @@
-package org.mobilenativefoundation.store.superstore5
-
-sealed class SuperstoreResponseOrigin {
-    object Cache : SuperstoreResponseOrigin()
-    object SourceOfTruth : SuperstoreResponseOrigin()
-    object Fetcher : SuperstoreResponseOrigin()
-    data class Warehouse<T : Any>(val value: T) : SuperstoreResponseOrigin()
-}
