--- superstore/src/commonMain/kotlin/org/mobilenativefoundation/store/superstore5/Warehouse.kt
@@ -1,9 +1,0 @@
-package org.mobilenativefoundation.store.superstore5
-
-/**
- * Represents a fallback data source.
- */
-interface Warehouse<Key : Any, out Output : Any> {
-    val name: String
-    suspend fun get(key: Key): WarehouseResponse<Output>
-}
