--- superstore/src/commonMain/kotlin/org/mobilenativefoundation/store/superstore5/WarehouseResponse.kt
@@ -1,10 +1,0 @@
-package org.mobilenativefoundation.store.superstore5
-
-sealed class WarehouseResponse<out Output : Any> {
-    data class Data<Output : Any>(
-        val data: Output,
-        val origin: String
-    ) : WarehouseResponse<Output>()
-
-    object Empty : WarehouseResponse<Nothing>()
-}
