--- superstore/src/commonTest/kotlin/org/mobilenativefoundation/store/superstore5/util/TestWarehouse.kt
@@ -1,5 +1,0 @@
-package org.mobilenativefoundation.store.superstore5.util
-
-import org.mobilenativefoundation.store.superstore5.Warehouse
-
-interface TestWarehouse<Key : Any, Output : Any> : Warehouse<Key, Output>
